requirements.txt
django
gunicorn
django-heroku
redis
rauth
python-dateutil
rq
